refactor: move space title editing logic to Spaces
